Adjust parameters for training environment and agent
Various hard-coded parameters in the training script were adjusted in order to improve stability of the training process and influence the learning performance of the agent. This includes increasing the prioritized replay buffer alpha, decreasing the memory size, increasing image fps, adjusting the environment max steps, and the learning rate. These changes will likely affect the agent's exploration, learning rate, and the overall speed of the training process.