Add optional rotation in image pre-processing
Image pre-processing function, preprocess_observation, has been updated to include an optional rotation parameter. This allows images to be randomly rotated during pre-processing if the parameter is activated, but will skip the step if it isn't. The function is used during memory buffer population, where the change is expected to add extra variability to training data.