Add option to include random door key while generating maze
Introduced a new parameter `add_random_door_key` to the `CustomEnv` class, allowing for optional inclusion of a random door key when generating a maze. If a predefined map is used, a warning is issued indicating that the random door key will not be added. Updated related methods to support this new parameter.